+++ src/sys/setup.hpp
@@ -89,7 +89,7 @@ class Setup {
   static constexpr int64_t DEFAULT_OEIS_UPDATE_INTERVAL =
       30;                                                 // 1 month default
   static constexpr int64_t DEFAULT_MAX_PROGRAM_AGE = 14;  // 2 weeks default
-  static constexpr int64_t DEFAULT_MAX_PHYSICAL_MEMORY = 1229;  // 1.2 GiB
+  static constexpr int64_t DEFAULT_MAX_PHYSICAL_MEMORY = 1296;  // 1.2 GiB
 
   static std::string LODA_HOME;
   static std::string SEQS_HOME;